requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
langchain-text-splitters
langchain-classic
langchain-core
pypdf
python-dotenv
sentence-transformers